Add support for flags per card.
Starting with being able to skip a card, or creating two cards with one item.
Some cleanup with how the heading was handled so that the flags are processed in
the same way as other cards -- which are now under a single mapping.  The counts
and output were updated to reflect the changes in data layout.

Removed the print_summary method since it no longer makes sense.