Fixed template precedence and updated option templ
Found that a card specific template choice was not always being preferred over
a section level choice.  Updated proc_flags to ensure that the ordering of flags
is from environment variables, sections, and then card specific.  This allows
check_template_flag to reliably choose the last flag related to its template.

Also updated the option template so that if the back of the card becomes the
front, there is some context for what is beeing looked for.