Prettify the output and add section comletion.
Added a number_space function to help pad numbers with spaces.  This helps the
output look cleaner.

Added an attribute to the objectives file that lets us tag a section as
complete.  This will alow for showing overall completion from two different
perspectives -- items in the checklist, and completion of sections (or topics if
preferred).