Handle path change for include statements.
The relative path needs to change when building the asciidoc examples file for a
given directory.  If there is a scripts directory, it is copied over, but it is
copied to the same directory as the new file.

This is a bit of a hack for now.  Should be handled better if the scripts ever
get consolidated in to a single tool.