Simplified the run_cmds script.
Shell expansion of the "line" variables was troublesome, and using variables
that should change behaviour had to be handled as an exception.  Also found
that trying to work with ulimit only applies to the current shell, breaking
the intent of the script.

Now, the command file is broken up (one command to be run per temporary file).
This allows for handling wierd errors that still occur seperately, and maintain
a flow around the command that created the output.