Added support for included examples.
Now that run_cmds can create Asciidoc snippets for the examples, it seems to
make sense to keep them in a seperate directory.  This change allows the numbers
for the examples to properly reflect those that come from these snippet files.

Note that the snippets in the 'includes' directory should not have an '.adoc'
extention since they are not full files, and we do not want them to be processed
directly.

My current convention is to have the "list of commands" file passed to run_cmds
be named 'cmds_*' while the output file is named 'inc_*'.  I seen no reason to
bake these in to the code.