Clean up for when errors occur.
When a command errors, the output is very different when sourced when compared
to executed directly on the shell.  Previously there was just a stub to detect
when this happened.  There is now an attempt to clean it up so that it looks
like it would if normally run on the CLI.