Improve run_cmds.
The run_cmds script now group commands together in a single literal block when
they are on contiguous lines in the input file.

It will also use an alternate string for the literal block if the previous
comments contain a '#:notexample' line.  This is so the example counts from
fcards do not include setup or teardown commands for the examples.