Minor fix for short files.
Found that short files were not working properly.  This was due to 'seq -w'
padding not padding zeros if there were less than 10 lines.  From the man page:

  equalize width by padding with leading zeroes

Of course it would not work.  Added a loop to update the NUM array if there are
less than 10 lines. Maybe not the best fix, but it will work for now.