Added option to collect examples in one file.
The new option for update_deck (-ce) required two more options in fcards
(-ce and -he).  When called with the new option, snippets are created for each
directory and and a header.  These are combined after the directories are
processed and built with the same Asciidoctor tool.

The subsections in examples will automatically be modified so that they work
as expected in the snippets.  No modification should be needed whether a
collected example file or individual files for each directory are needed.