Fixed update when included files changed.
Found that trying to keep all of the built files in one directory was more
trouble than it was worth.  Decided that renamine the output directory and using
a recursive diff would be much easier.  It is a bit messy to have multiple
exclude patterns for the diff.  Perhaps moving the cards to a seperate directory
would be cleaner.

The command files (cmd_*) really shouldn't be baked in here either.  It implies
that the directory structure is predetermined and its not intended to be.  They
are only files that are used to generate the command examples with run_cmds.

Things to improve at a later date.